requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers>=4.44.0
sentencepiece>=0.2.0

# App
streamlit>=1.37.0
Pillow>=10.3.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.2
rapidfuzz>=3.9.6
